Standardising the validity checks for entity type mapping
Now there is a single implementation that does checks to see whether a type meets the criteria to be mapped.

Additionally, this new standardised version enforces that types that are assignable to BsonValue are not valid for mapping their own entity definition. This is because these types themselves represent the raw serializable type - no entity mapping necessary.