Works around issues with the ResultTransformer and query generation
MongoDB driver internally uses the ResultTransformer to do after-query LINQ stuff. For `Any()`, it makes the whole entity come back from the DB and then checks whether anything came back. This is a problem for entity processing system as it sees a full entity and triggers the entity tracking processor, potentially squashing any entities in there for no reason.

Solution is to check if the ResultTransformer is set and if the return type matches the entity type.